Add the structure for versions in CustomResourceDefinition.
This is the first part for CustomResourceDefinition migration.
The change includes re-architect the BottlerocketShadow in
models/src/node, also adjusts the way to generate CRD's yaml file.
Right now webhook is not suported yet, the default version
conversion is None.
